--- a/launch/tier4_planning_launch/launch/scenario_planning/parking.launch.xml
+++ b/launch/tier4_planning_launch/launch/scenario_planning/parking.launch.xml
@@ -3,7 +3,8 @@
   <group if="$(var launch_parking_module)">
     <push-ros-namespace namespace="parking"/>
 
-      <node pkg="autoware_costmap_generator" exec="costmap_generator" name="costmap_generator" namespace="" launch-prefix="konsole -e gdb -ex run --args">
+    <node_container pkg="rclcpp_components" exec="$(var container_type)" name="parking_container" namespace="" output="screen">
+      <composable_node pkg="autoware_costmap_generator" plugin="autoware::costmap_generator::CostmapGenerator" name="costmap_generator" namespace="">
         <!-- topic remap -->
         <remap from="~/input/objects" to="/perception/object_recognition/objects"/>
         <remap from="~/input/points_no_ground" to="/perception/obstacle_segmentation/pointcloud"/>
@@ -14,10 +15,9 @@
         <!-- params -->
         <param from="$(var vehicle_param_file)"/>
         <param from="$(var costmap_generator_param_path)"/>
-        <!-- <extra_arg name="use_intra_process_comms" value="false"/> -->
-      </node>
+        <extra_arg name="use_intra_process_comms" value="false"/>
+      </composable_node>
 
-    <node_container pkg="rclcpp_components" exec="$(var container_type)" name="parking_container" namespace="" output="screen">
       <composable_node pkg="autoware_freespace_planner" plugin="autoware::freespace_planner::FreespacePlannerNode" name="freespace_planner" namespace="">
         <!-- topic remap -->
         <remap from="~/input/route" to="/planning/mission_planning/route"/>