--- a/planning/autoware_costmap_generator/include/autoware_costmap_generator/objects_to_costmap.hpp
+++ b/planning/autoware_costmap_generator/include/autoware_costmap_generator/objects_to_costmap.hpp
@@ -103,7 +103,7 @@ private:
   /// \param[in] in_corner_point one of convex hull points
   /// \param[in] expand_polygon_size  the param for expanding convex_hull points
   /// \param[out] expanded point
-  geometry_msgs::msg::Point makeExpandedPoint(
+  static geometry_msgs::msg::Point makeExpandedPoint(
     const geometry_msgs::msg::Point & in_centroid,
     const geometry_msgs::msg::Point32 & in_corner_point, const double expand_polygon_size);
 
@@ -111,7 +111,7 @@ private:
   /// \param[in] in_centroid: object's centroid
   /// \param[in] expand_polygon_size: expanding convex_hull points
   /// \param[out] polygon object with convex hull points
-  grid_map::Polygon makePolygonFromObjectConvexHull(
+  static grid_map::Polygon makePolygonFromObjectConvexHull(
     const std_msgs::msg::Header & header,
     const autoware_perception_msgs::msg::PredictedObject & in_object,
     const double expand_polygon_size);
@@ -121,7 +121,7 @@ private:
   /// \param[in] gridmap_layer_name: target gridmap layer name for calculated cost
   /// \param[in] score: set score as a cost for costmap
   /// \param[in] objects_costmap: update cost in this objects_costmap[gridmap_layer_name]
-  void setCostInPolygon(
+  static void setCostInPolygon(
     const grid_map::Polygon & polygon, const std::string & gridmap_layer_name, const float score,
     grid_map::GridMap & objects_costmap);
 };